# ci: set up ccache for macOS builds