--- a/.ccache/ccache.conf
+++ b/.ccache/ccache.conf
@@ -1,0 +1,5 @@
+depend_mode = true
+direct_mode = true
+hard_link = true
+run_second_cpp = true
+sloppiness = file_stat_matches,include_file_ctime,include_file_mtime,ivfsoverlay,modules,system_headers,time_macros