--- a/scripts/macos_e2e.sh
+++ b/scripts/macos_e2e.sh
@@ -2,6 +2,23 @@
 
 BUILD_ACTIONS="$@"
 
+if [[ "$CCACHE_DISABLE" != "1" ]]; then
+  if ! command -v ccache 1> /dev/null; then
+    brew install ccache
+  fi
+
+  CCACHE_HOME=$(dirname $(dirname $(which ccache)))/opt/ccache
+
+  export CCACHE_DIR="$(git rev-parse --show-toplevel)/.ccache"
+
+  export CC="${CCACHE_HOME}/libexec/clang"
+  export CXX="${CCACHE_HOME}/libexec/clang++"
+  export CMAKE_C_COMPILER_LAUNCHER=$(which ccache)
+  export CMAKE_CXX_COMPILER_LAUNCHER=$(which ccache)
+
+  ccache --zero-stats 1> /dev/null
+fi
+
 # Workaround for `Element StaticText, {{163.0, 836.0}, {156.0, 21.0}}, value: Set native delegate is not hittable`.
 # This occurs when a button is not visible in the window. We resize the window
 # here to ensure that it is.
@@ -14,5 +31,9 @@ xcodebuild \
   -sdk macosx \
   -derivedDataPath example/macos/build \
   $BUILD_ACTIONS
+
+if [[ "$CCACHE_DISABLE" != "1" ]]; then
+  ccache --show-stats --verbose
+fi
 
 exit $?