## Summary

Set up ccache for macOS builds.

## Test Plan

CI should pass. First builds should take as much time as it does today, but subsequent ones should be significantly faster.

| Before | After |
| - | - |
| [24m 32s](https://github.com/react-native-async-storage/async-storage/runs/8203935502?check_suite_focus=true) | [13m 31s](https://github.com/react-native-async-storage/async-storage/runs/8204398679?check_suite_focus=true) |