@krizzu 2022-11-11
:tada: This PR is included in version 1.17.11 :tada:

The release is available on:
- [npm package (@latest dist-tag)](https://www.npmjs.com/package/@react-native-async-storage/async-storage/v/1.17.11)
- [GitHub release](https://github.com/react-native-async-storage/async-storage/releases/tag/v1.17.11)

Your **[semantic-release](https://github.com/semantic-release/semantic-release)** bot :package::rocket: